move path resolution from bundlers config to dev-tools
