+++ apps/app-esbuild/config/build.config.ts
@@ -1,30 +1,14 @@
 import * as esbuild from "esbuild";
 import { getEnvVariables } from "@repo/dev-tools/config/environment";
-import path from "path";
-
-import tsNodeLoader from 'ts-node';
-
-tsNodeLoader.register({
-  esm: true
-});
+import "@repo/dev-tools/config/paths";
+import { appCoreEnvDir } from "@repo/dev-tools/config/paths";
 
 const BUILD_DIR = "dist";
 
-console.log("");
-console.log("* process.env.NODE_ENV:", process.env.NODE_ENV);
-console.log("* process.env.BUILD_ENVIRONMENT", process.env.BUILD_ENVIRONMENT);
-console.log("");
-
-const __dirname = process.cwd();
-const pathResolve = (pathEntry: string) => path.resolve(__dirname, pathEntry);
-// const appCorePublic = pathResolve("./node_modules/@repo/app-core/public");
-// const appCoreEnv = pathResolve("./node_modules/@repo/app-core/.env");
-const appCoreEnvDir = pathResolve("./node_modules/@repo/app-core/");
-
 const environmentVariables = getEnvVariables(
   appCoreEnvDir,
-  process.env.BUILD_ENVIRONMENT ?? 'production',
-  "esbuild",
+  process.env.BUILD_ENVIRONMENT ?? "production",
+  "esbuild"
 );
 
 async function bundle() {
@@ -52,7 +36,7 @@ async function bundle() {
     console.log("Build successfull");
     // console.log(buildResult);
   } catch (err) {
-    console.log('Error during build:', err);
+    console.log("Error during build:", err);
   }
 }
